Weather repositories logic changes
And other minor fixes

+++ app/src/main/kotlin/ru/mts/avpopo85/weathery/data/model/base/openWeatherMap/currentWeather/IOWMCurrentWeatherMainResponse.kt
@@ -10,7 +10,7 @@ interface IOWMCurrentWeatherMainResponse {
     val atmosphericPressureInhPa: Double
 
     /**Humidity, %*/
-    val humidity: Int
+    val humidity: Double
 
     /**Minimum temperature at the moment.
      *
