+++ app/src/main/kotlin/ru/mts/avpopo85/weathery/data/model/implementation/openWeatherMap/forecast/OWMForecastResponse.kt
@@ -12,7 +12,7 @@ data class OWMForecastResponse(
 
     @SerializedName("cnt")
     @Expose
-    override val numberOfLines: Double,
+    override val numberOfLines: Int,
 
     @SerializedName("cod")
     @Expose
